+++ src/components/Header/Header.tsx
@@ -1,3 +1,4 @@
+import React from "react";
 import Button from "../Button/Button";
 
 const Header = () => {
