+++ src/index.js
@@ -3,7 +3,7 @@ import ReactDOM from "react-dom";
 import "./index.scss";
 import App from "./App";
 import reportWebVitals from "./reportWebVitals";
-import store from "./redux/store";
+import { store } from "./redux/store";
 import { Provider } from "react-redux";
 
 ReactDOM.render(
