+++ src/redux/store.ts
@@ -1,8 +1,11 @@
 import { configureStore } from "@reduxjs/toolkit";
 import gifListReducer from "./slices/gifSlice";
 
-export default configureStore({
+export const store = configureStore({
   reducer: {
     gifList: gifListReducer,
   },
 });
+
+export type AppDispatch = typeof store.dispatch;
+export type RootState = ReturnType<typeof store.getState>;
